Isolate jar extension
